@bas suggestion to read and re-set

+++ scripts/reset-demo.sh
@@ -13,17 +13,41 @@ echo "Re-setting Demo"
 
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 
-# Disabling protected branches otherwise force push fails (when available in Enterprise)
-curl -H "Authorization: Token $GITHUB_TOKEN" -H "Accept: application/vnd.github.loki-preview+json" -H "Content-type: application/json" -X DELETE https://octodemo.com/api/v3/repos/$RT_ORG/$RT_REPO/branches/master/protection
+# Check if master branch is protected
+BRANCHES_STATUS_CODE=$(curl -I -H "Authorization: Token $GITHUB_TOKEN" -H "Accept: application/vnd.github.loki-preview+json" https://octodemo.com/api/v3/repos/$RT_ORG/$RT_REPO/branches/master/protection -s -o /dev/null -w %{http_code})
 
+if [ $BRANCHES_STATUS_CODE -ne "404" ] ; then
+
+    echo "Master is protected: disabling protection"
+
+    # Read required status currently activated on master branch
+    CONTEXTS=$(curl -H "Authorization: Token $GITHUB_TOKEN" -H "Accept: application/vnd.github.loki-preview+json" https://octodemo.com/api/v3/repos/$RT_ORG/$RT_REPO/branches/master/protection/required_status_checks/contexts)
+
+    # Remove new lines from $CONTEXTS
+    CONTEXTS=$(echo $CONTEXTS|tr -d '\n')
+
+    # Disabling protected branches otherwise force push fails (when available in Enterprise)
+    curl -H "Authorization: Token $GITHUB_TOKEN" -H "Accept: application/vnd.github.loki-preview+json" -H "Content-type: application/json" -X DELETE https://octodemo.com/api/v3/repos/$RT_ORG/$RT_REPO/branches/master/protection
+
+else
+  echo "Master is not protected"
+fi
+
+# Force push HEAD to baseline
 echo "Reverting master to baseline tag"
 git fetch --tags
 git checkout master
 git reset --hard baseline
 git push origin baseline:master -f
 
-# Re-enabling protected branches / required status
-curl -H "Authorization: Token $GITHUB_TOKEN" -H "Accept: application/vnd.github.loki-preview+json" -H "Content-type: application/json" -X PUT https://octodemo.com/api/v3/repos/$RT_ORG/$RT_REPO/branches/master/protection -d @protected.json
+if [ "$BRANCHES_STATUS_CODE" -ne 404 ] ; then
+
+    echo "Re-enabling protected branches as before: $CONTEXTS"
+
+    # Get the new JSON based on the protected.json template and run the re-enable API
+    sed -e "s|CONTEXTS_PLACEHOLDER|$CONTEXTS|g" scripts/protected.json | curl -H "Authorization: Token $GITHUB_TOKEN" -H "Accept: application/vnd.github.loki-preview+json" -H "Content-type: application/json" -X PUT https://octodemo.com/api/v3/repos/$RT_ORG/$RT_REPO/branches/master/protection -d @-
+fi
+
 
 
 function close_issue () {
